test using react native built in picker.
Reason for test is because when app is uploaded to snack, snack gives error that the react native picker library is not supported and must be updated to an earlier version. According to the documentation, use of react native in-built picker component is necessary for managed workflow of expo sdk37.